order shop update

+++ order-service/src/main/java/com/hoangtien2k3/orderservice/entity/Cart.java
@@ -25,7 +25,7 @@ public final class Cart extends AbstractMappedEntity {
     private Integer cartId;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId; // userId
 
     @JsonIgnore
     @OneToMany(mappedBy = "cart", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
